+++ src/pages/CityList/Index.js
@@ -2,10 +2,11 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Jumbotron } from "react-bootstrap";
 import Button from 'react-bootstrap/Button';
-import { fetchCityList, fetchCityByCondition } from '../../store/cityList/actions';
+import { fetchCityList } from '../../store/cityList/actions';
 import { selectCities } from '../../store/cityList/selectors';
 import CityCard from './CityCard';
 import { selectFilteredCities } from '../../store/cityFilter/selectors';
+import { fetchCityByCondition } from '../../store/cityFilter/actions';
 
 export default function CityList() {
   const dispatch = useDispatch();
@@ -42,12 +43,6 @@ export default function CityList() {
 
 
   const clickHandler = () => {
-    console.log(`
-      continent: ${continent}
-      country: ${country}
-      population: ${population}
-      price: ${price}
-    `)
     dispatch(fetchCityByCondition(country, continent, population, price));
     setToggle('filter');
   }
@@ -61,30 +56,41 @@ export default function CityList() {
       </Jumbotron>
 
       <div className='filter-buttons'>
-        <label htmlFor="country-select">Sort by country:</label>
-        <select name="country" id="country-select" onChange={(e) => setCountry(e.target.value)}>
-          <option value="all">--Please choose an option--</option>
-          <option value="all">all</option>
-           {citiesList.map(city => {
-             return (
-                <option value={city.country} key={city.id}>{city.country}</option>
-             )
-           })}
-        </select>
-        <label htmlFor="continent-select">Sort by continent:</label>
-        <select name="continent" id="continent-select" onChange={(e) => setContinent(e.target.value)}>
-          <option value="all">--Please choose an option--</option>
-          <option value="all">all</option>
-           {citiesList.map(city => {
-             return (
-                <option value={city.continent} key={city.id}>{city.continent}</option>
-             )
-           })}
-        </select>
-        <label>Minimum value of population:</label>
-        <input type="range" value={population} min={minPopulation} max={maxPopulation} onChange={e => setPopulation(e.target.value)} />
-        <label>Maximum price:</label>
-        <input type="range" value={price} min={minPrice} max={maxPrice} onChange={e => setPrice(e.target.value)} />
+        <div className='country-filter'>
+          <label htmlFor="country-select">Sort by country:</label>
+          <select name="country" id="country-select" onChange={(e) => setCountry(e.target.value)}>
+            <option value="all">--Please choose an option--</option>
+            <option value="all">all</option>
+            {citiesList.map(city => {
+              return (
+                  <option value={city.country} key={city.id}>{city.country}</option>
+              )
+            })}
+          </select>
+        </div>
+
+        <div className='continent-filter'>
+          <label htmlFor="continent-select">Sort by continent:</label>
+          <select name="continent" id="continent-select" onChange={(e) => setContinent(e.target.value)}>
+            <option value="all">--Please choose an option--</option>
+            <option value="all">all</option>
+            {citiesList.map(city => {
+              return (
+                  <option value={city.continent} key={city.id}>{city.continent}</option>
+              )
+            })}
+          </select>
+        </div>
+        
+        <div className='population-filter'>
+          <label>Minimum value of population:</label>
+          <input type="range" value={population} min={minPopulation} max={maxPopulation} onChange={e => setPopulation(e.target.value)} />
+        </div>
+
+        <div className='price-filter'>
+          <label>Maximum price:</label>
+          <input type="range" value={price} min={minPrice} max={maxPrice} onChange={e => setPrice(e.target.value)} />
+        </div>
 
         <Button variant="primary" onClick={clickHandler}>Set filters</Button>
         <Button variant="primary" onClick={() => setToggle('list')}>Reset</Button>
