+++ src/store/cityList/actions.js
@@ -10,15 +10,3 @@ export function fetchCityList() {
   }
 }
 
-export function fetchCityByCondition(country, continent, population, price) {
-  return async (dispatch, getState) => {
-    try {
-      const response = await axios.get(`${apiUrl}/cities/${country}/${continent}/${population}/${price}`);
-      
-      dispatch({ type: 'FILTER_CITIES', payload: response.data })
-    }
-    catch (error) {
-      console.log('error: ', error);
-    }
-  }
-}
